сделал точечный свет и повесил его себе на лоб.

+++ PointLight.cpp
@@ -1,9 +1,9 @@
 #include "PointLight.h"
 
 
-void PointLight::UseLight(GLfloat ambientIntnsityLocation, GLfloat ambientColourLocation, 
-						  GLfloat diffuseIntensityLocation, GLfloat positionLocation, 
-	                      GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntnsityLocation, GLfloat ambientColourLocation,
+						  GLuint diffuseIntensityLocation, GLuint positionLocation,
+						  GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
 	glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
 	glUniform1f(ambientIntnsityLocation, ambientIntensity);
